Created init file and brought unittest file inside the main folder
